:truck: api: move find-user-private-certificates usecase from lib to src
