+++ api/src/certification/results/domain/usecases/find-user-private-certificates.js
